// NetscriptDefinitions.d.ts
export interface GangOtherInfoObject {
    /** Gang power */
    power: number;
    /** Gang territory, in decimal form */
    territory: number;
}
// ...
export interface GangTaskStats {
    /** Task name */
    name: string;
    /** Task Description */
    desc: string;
    /** Is a task of a hacking gang */
    isHacking: boolean;
    /** Is a task of a combat gang */
    isCombat: boolean;
    /** Base respect earned */
    baseRespect: number;
    /** Base wanted earned */
    baseWanted: number;
    /** Base money earned */
    baseMoney: number;
    /** Hacking skill impact on task scaling */
    hackWeight: number;
    /** Strength skill impact on task scaling */
    strWeight: number;
    /** Defense skill impact on task scaling */
    defWeight: number;
    /** Dexterity skill impact on task scaling */
    dexWeight: number;
    /** Agility skill impact on task scaling */
    agiWeight: number;
    /** Charisma skill impact on task scaling */
    chaWeight: number;
    /** Number representing the difficulty of the task */
    difficulty: number;
    /** Territory impact on task scaling */
    territory: GangTerritory;
}
// ...
export interface GangTerritory {
    /** Money gain impact on task scaling */
    money: number;
    /** Respect gain impact on task scaling */
    respect: number;
    /** Wanted gain impact on task scaling */
    wanted: number;
}

/**
 * @public
 */
export interface GangMemberInfo {
    /** Name of the gang member */
    name: string;
    /** Currently assigned task */
    task: string;
    earnedRespect: number;

    /** Hack skill level */
    hack: number;
    /** Strength skill level */
    str: number;
    /** Defense skill level */
    def: number;
    /** Dexterity skill level */
    dex: number;
    /** Agility skill level */
    agi: number;
    /** Charisma skill level */
    cha: number;

    /** Current hack experience */
    hack_exp: number;
    /** Current strength experience */
    str_exp: number;
    /** Current defense experience */
    def_exp: number;
    /** Current dexterity experience */
    dex_exp: number;
    /** Current agility experience */
    agi_exp: number;
    /** Current charisma experience */
    cha_exp: number;

    /** Hack multiplier from equipment */
    hack_mult: number;
    /** Strength multiplier from equipment */
    str_mult: number;
    /** Defense multiplier from equipment */
    def_mult: number;
    /** Dexterity multiplier from equipment */
    dex_mult: number;
    /** Agility multiplier from equipment */
    agi_mult: number;
    /** Charisma multiplier from equipment */
    cha_mult: number;

    /** Hack multiplier from ascensions */
    hack_asc_mult: number;
    /** Strength multiplier from ascensions */
    str_asc_mult: number;
    /** Defense multiplier from ascensions */
    def_asc_mult: number;
    /** Dexterity multiplier from ascensions */
    dex_asc_mult: number;
    /** Agility multiplier from ascensions */
    agi_asc_mult: number;
    /** Charisma multiplier from ascensions */
    cha_asc_mult: number;

    /** Total earned hack experience */
    hack_asc_points: number;
    /** Total earned strength experience */
    str_asc_points: number;
    /** Total earned defense experience */
    def_asc_points: number;
    /** Total earned dexterity experience */
    dex_asc_points: number;
    /** Total earned agility experience */
    agi_asc_points: number;
    /** Total earned charisma experience */
    cha_asc_points: number;

    upgrades: string[];
    augmentations: string[];

    respectGain: number;
    wantedLevelGain: number;
    moneyGain: number;
}